refactor: Split templates.jsx into multiple files.
