+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 import { Routes, Route } from "react-router-dom";
 import Dashboard from "./components/dashboard/dashboard";
-import Template from "./components/templates";
+import Template from "./components/templates/templates";
 import NotFound from "./components/notfound";
 import Accounts from "./components/accounts";
 import Profile from "./components/profile";
@@ -12,22 +12,20 @@ import Register from "./components/authentication/register";
 
 function App() {
     return (
-        <>
-            <GlobalProvider>
-                <Navbar />
-                <Routes>
-                    <Route path='/dashboard' element={<Dashboard />} />
-                    <Route path='/template' element={<Template />} />
-                    <Route path='/' element={<Dashboard />} />
-                    <Route path='/accounts' element={<Accounts />} />
-                    <Route path='/profile' element={<Profile />} />
-                    <Route path='/login' element={<Login />} />
-                    <Route path='/register' element={<Register />} />
-                    <Route path='' element={<Register />} />
-                    <Route path='*' element={<NotFound />} />
-                </Routes>
-            </GlobalProvider>
-        </>
+        <GlobalProvider>
+            <Navbar />
+            <Routes>
+                <Route path='/dashboard' element={<Dashboard />} />
+                <Route path='/template' element={<Template />} />
+                <Route path='/' element={<Dashboard />} />
+                <Route path='/accounts' element={<Accounts />} />
+                <Route path='/profile' element={<Profile />} />
+                <Route path='/login' element={<Login />} />
+                <Route path='/register' element={<Register />} />
+                <Route path='' element={<Register />} />
+                <Route path='*' element={<NotFound />} />
+            </Routes>
+        </GlobalProvider>
     );
 }
