+++ frontend/src/components/templates/templates.scss
@@ -1,4 +1,4 @@
-@import "../mixins";
+@import "../../mixins";
 
 .template-wrapper {
     margin-top: 80px;
